kiaf_farmer_backend/requirements.txt
asgiref==3.3.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-admin-honeypot==1.1.0
django-cors-headers==3.6.0
django-ipware==3.0.2
django-mssql-backend==2.8.1
django-otp==1.0.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-recaptcha==2.0.4
drf-yasg==1.20.0
facebook-sdk==3.1.0
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
micropython-fcntl==0.0.3
micropython-ffilib==0.1.2
oauthlib==3.1.0
packaging==20.8
protobuf==3.17.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.0
pyodbc==4.0.30
pyparsing==2.4.7
python-decouple==3.4
python-twitter==3.5
pytz==2020.5
qrcode==6.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
